Move LogStatistics to util
